alpha blending now works

+++ ProjectDibenaySDL/Province.h
@@ -9,14 +9,14 @@ namespace Mer
 	{
 	public:
 		Province();
-		Province(std::string id, int red, int green, int  blue, int screenWidth, int ScreenHeight);
+		Province(std::string id, Uint8 red, Uint8 green, Uint8  blue, int screenWidth, int ScreenHeight);
 
-		int getRed();
-		int getGreen();
-		int getBlue();
+		Uint8 getRed();
+		Uint8 getGreen();
+		Uint8 getBlue();
 		std::string getID();
 
-		void MapPixelColour(int x, int y, int red, int green, int blue, int alpha);
+		void MapPixelColour(int x, int y, Uint8 red, Uint8 green, Uint8 blue, Uint8 alpha);
 
 		SDL_Texture* getTexture();
 		SDL_Surface* getSurface();
@@ -31,9 +31,9 @@ namespace Mer
 		Uint32* pixels;
 		SDL_Texture* provinceTexture;
 		std::string id = "";
-		int red = 0;
-		int green = 0;
-		int blue = 0;
+		Uint8 red = 0;
+		Uint8 green = 0;
+		Uint8 blue = 0;
 
 		SDL_Rect sRect;
 		SDL_Rect dRect;
